move sql connection to separate file
