+++ bamazonCustomer.js
@@ -1,23 +1,8 @@
 var inquirer = require("inquirer");
-var mysql = require("mysql");
+var connection = require("./connection.js")
 
 
-var connection = mysql.createConnection({
-    host: "localhost",
-    port: 3306,
-    user: "root",
-    password: "",
-    database: "bamazon_db"
-});
-
-
-connection.connect(function(error) {
-    if (error) throw error;
-    console.log("connected as id " + connection.threadId);
-    console.log("I have connected to mysql!");
-    runCustomer();
-});
-
+runCustomer();
 
 function runCustomer(query) {
     var productArr = [0];
