+++ bamazonSupervisor.js
@@ -1,21 +1,9 @@
 var inquirer = require("inquirer");
-var mysql = require("mysql");
-
-var connection = mysql.createConnection({
-    host: "localhost",
-    port: 3306,
-    user: "root",
-    password: "",
-    database: "bamazon_db"
-});
+var connection = require("./connection.js")
 
 
-connection.connect(function(error) {
-    if (error) throw error;
-    console.log("connected as id " + connection.threadId);
-    console.log("I have connected to mysql!");
-    runSupervisor();
-});
+runSupervisor();
+
 
 // TODO: Add department_ids on the fly
 function runSupervisor() {
